refactoring temporal profiles
