+++ emiproc/profiles/temporal/constants.py
@@ -1,0 +1,9 @@
+# Constants
+N_HOUR_DAY = 24
+N_DAY_WEEK = 7
+N_MONTH_YEAR = 12
+N_DAY_YEAR = 365
+N_DAY_LEAPYEAR = 366
+N_HOUR_WEEK = N_HOUR_DAY * N_DAY_WEEK
+N_HOUR_YEAR = N_DAY_YEAR * N_HOUR_DAY
+N_HOUR_LEAPYEAR = N_DAY_LEAPYEAR * N_HOUR_DAY
